refactor: move Dependency class to dependency.py
